% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CoefFunReg.r
\name{coef.funreg}
\alias{coef.funreg}
\title{coef method for funreg object}
\usage{
\method{coef}{funreg}(object, digits = 4, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{funreg}}

\item{digits}{The number of digits past the decimal place to use when printing numbers}

\item{silent}{If \code{TRUE}, indicates that the summary should be returned
as a list object but not printed to the screen.}

\item{...}{Other arguments that may be passed from another method.}
}
\value{
At least for now, this is identical to the \code{summary.funreg}
function.
}
\description{
Returns coefficient information on a \code{funreg} object.
}

