% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NumFunctionalCovsInModel.r
\name{num.functional.covs.in.model}
\alias{num.functional.covs.in.model}
\title{Count the functional covariates in a model (for internal use by other package functions)}
\usage{
num.functional.covs.in.model(object)
}
\arguments{
\item{object}{An object of class \code{funreg}, representing
a fitted penalized functional regression with one or more
functional covariates.}
}
\value{
The number of functional covariates as an integer.
}
\description{
A very simple function, mainly for internal use by
package code, to count the number of functional
covariates in an object of class \code{funreg}.
}

