% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotFunEigen.r
\name{plot.funeigen}
\alias{plot.funeigen}
\title{plot method for funeigen object}
\usage{
\method{plot}{funeigen}(x, type = "correlation", how.many = NULL,
  xlab = "", ylab = "", ...)
}
\arguments{
\item{x}{An object of class \code{funeigen}}

\item{type}{A character string telling the kind
of information to include in the plot.  It may be \code{functions},
\code{eigenfunctions}, \code{eigenvalues}, \code{mean}, \code{covariance}, or \code{correlation}.}

\item{how.many}{How many fitted curves to show (in a plot of
fitted curves; the default is all of them), or how many estimated
eigenfunctions to show (in a plot of eigenfunctions; the default is
all of them)}

\item{xlab}{Label for the x axis of the plot.}

\item{ylab}{Label for the y axis of the plot.}

\item{...}{Other optional arguments to be passed on to the plot function.}
}
\description{
Creates a visual representation of some of the
information in an object of class \code{funeigen}
(i.e., in an eigenfunction decomposition of a
functional variable).  Several kinds of plots are
available.
}

