% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotFunReg.r
\name{plot.funreg}
\alias{plot.funreg}
\title{plot method for funreg object}
\usage{
\method{plot}{funreg}(x, frames = FALSE, type = "coefficients", ...)
}
\arguments{
\item{x}{An object of class \code{funreg}, representing
a fitted functional regression model.}

\item{frames}{If there are multiple functional covariates, this
tells whether or not the plot for each covariate should all be drawn together as different
panels on the same figure.}

\item{type}{A string telling what kind of plot to produce.
One can specify \code{coefficients}, which means that the
functional coefficient for each functional covariate will be
plotted as a function of time; or one can specify}

\item{...}{Other optional arguments to be passed on to the plot function.
 \code{correlations}, which means that the correlation
  of the covariate (at each given time), with
the scalar outcome (presumably taken at a single
 time) will be plotted instead.}
}
\description{
Plots information from an object of class \code{funreg}.
}

