% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PrintFunReg.r
\name{print.funreg}
\alias{print.funreg}
\title{print method for funreg object}
\usage{
\method{print}{funreg}(x, digits = 4, show.fits = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{funreg}.}

\item{digits}{Number of digits past the decimal place
to show when printing quantities.}

\item{show.fits}{Whether to also print a table of fitted values for
the functional coefficients along a grid of times.}

\item{...}{Any other optional arguments to be passed to the default print
function.}
}
\description{
Prints information from an object of class \code{funreg}.
}

