% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RedoFunReg.r
\name{redo.funreg}
\alias{redo.funreg}
\title{Redo a funreg with different data (for internal use by permutation test)}
\usage{
redo.funreg(object, id, response, time, other.covariates, x)
}
\arguments{
\item{object}{A \code{funreg} object.}

\item{id}{The new values for id (see the \code{funreg}
function documentation)}

\item{response}{The new values for response}

\item{time}{The new values for time}

\item{other.covariates}{The new values for other.covariates (which may be NULL)}

\item{x}{The new values for the functional covariate.}
}
\value{
The \code{funreg} object for the new fitted model.
}
\description{
For internal use by package functions.  Not
 intended to be directly called by the data analyst.
 This function repeats the analysis for an object
 of class \code{funreg}, with the same settings
 but with possibly different data.  This is
 convenient in doing resampling techniques like
 bootstrapping or permutation testing.
}

