% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SummaryFunReg.r
\name{summary.funreg}
\alias{summary.funreg}
\title{summary method for funreg object}
\usage{
\method{summary}{funreg}(object, digits = 4, silent = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{funreg}}

\item{digits}{The number of digits past the decimal place to use when printing numbers}

\item{silent}{If \code{TRUE}, indicates that the summary should be returned
as a list object but not printed to the screen.}

\item{...}{Any other optional arguments that may be passed
from other methods (but currently ignored by this one).}
}
\value{
Returns a list with four components.  First, \code{call.info} summarizes the
inputs that were sent into the \code{funreg} function.  Second,
\code{intercept.estimate.uncentered} gives the estimated functional
coefficient for the intercept in the functional regression model.  Third,
\code{functional.covariates.table} provides estimated values for the
functional coefficients at each of a grid of time points.  Fourth,
\code{subject.level.covariates.table} provides estimated values for
subject-level covariates if any are in the model.
}
\description{
Returns summary information on a \code{funreg} object.
}

